+++ package.json
@@ -23,7 +23,7 @@
     "@types/object-hash": "1.2.0",
     "ava": "0.25.0",
     "graphql": "0.13.2",
-    "graphql-middleware": "1.3.3",
+    "graphql-middleware": "^1.4.1",
     "graphql-tools": "3.0.5",
     "prettier": "1.13.7",
     "prettier-check": "2.0.0",
@@ -36,7 +36,7 @@
   },
   "peerDependencies": {
     "graphql": "^0.11.0 || ^0.12.0 || ^0.13.0",
-    "graphql-middleware": "^1.3.2"
+    "graphql-middleware": "^1.4.1"
   },
   "files": [
     "dist"
