+++ src/customError.ts
@@ -1,0 +1,5 @@
+export class CustomError extends Error {
+  constructor(...args) {
+    super(...args)
+  }
+}
