+++ src/fragmentReplacements.ts
@@ -1,0 +1,7 @@
+import { IRules, FragmentReplacement } from './types'
+
+export const extractFragmentReplacements = (
+  shield: IRules,
+): FragmentReplacement[] => {
+  return []
+}
