+++ tsconfig.json
@@ -9,5 +9,5 @@
     "lib": ["dom", "es2017", "esnext.asynciterable"],
     "experimentalDecorators": true
   },
-  "exclude": ["node_modules", "examples"]
+  "exclude": ["node_modules", "examples", "dist"]
 }
